# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      description: >
        Query the Task Manager's task history. Non-admin callers only see
        tasks scoped to their own account.
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: type
          in: query
          schema:
            type: string
        - name: queue
          in: query
          schema:
            type: string
        - name: correlationId
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Task page
          content:
            application/json:
              schema:
                type: object
                required: [tasks]
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/OctopusTask"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Enqueue a task
      description: >
        Creates a new task in the `queued` state and publishes a
        `task.created` event. The task is not executed by this call — a
        worker claims it via POST /tasks/claim.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: The created task
          content:
            application/json:
              schema:
                type: object
                required: [task]
                properties:
                  task:
                    $ref: "#/components/schemas/OctopusTask"
        "400":
          description: Invalid request body
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a single task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The task
          content:
            application/json:
              schema:
                type: object
                required: [task]
                properties:
                  task:
                    $ref: "#/components/schemas/OctopusTask"
        "404":
          description: Not found
  /tasks/{id}/cancel:
    post:
      operationId: cancelTask
      tags: [tasks]
      summary: Cancel a task
      description: >
        Cancels a task that is still `queued` or `running`, and publishes a
        `task.cancelled` event. No-op error if the task is already in a
        terminal state.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The cancelled task
          content:
            application/json:
              schema:
                type: object
                required: [task]
                properties:
                  task:
                    $ref: "#/components/schemas/OctopusTask"
        "404":
          description: Not found
        "409":
          description: Task is already in a terminal state
  /tasks/claim:
    post:
      operationId: claimTask
      tags: [tasks]
      summary: Claim the next available task
      description: >
        Queue Manager operation: atomically claims the next eligible queued
        task for the caller, marks it `running`, and publishes a
        `task.started` event. Intended for worker/agent clients. Returns 204
        (no body) if no task is currently eligible.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                queue:
                  type: string
                  description: Defaults to "default"
      responses:
        "200":
          description: The claimed task
          content:
            application/json:
              schema:
                type: object
                required: [task]
                properties:
                  task:
                    $ref: "#/components/schemas/OctopusTask"
        "204":
          description: No eligible task to claim
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a claimed task completed
      description: >
        Queue Manager operation: marks a task this caller currently holds
        the claim on as `completed`, and publishes a `task.completed` event.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                result: {}
      responses:
        "200":
          description: The completed task
          content:
            application/json:
              schema:
                type: object
                required: [task]
                properties:
                  task:
                    $ref: "#/components/schemas/OctopusTask"
        "404":
          description: Not found
        "409":
          description: Task is not currently claimed by this caller
  /tasks/{id}/fail:
    post:
      operationId: failTask
      tags: [tasks]
      summary: Report a claimed task as failed
      description: >
        Queue Manager operation: reports failure for a task this caller
        currently holds the claim on. If retries remain (and `retry` isn't
        `false`), reschedules the task with exponential backoff and
        publishes `task.retrying`; otherwise marks it permanently `failed`
        and publishes `task.failed`.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [error]
              properties:
                error:
                  type: string
                retry:
                  type: boolean
                  description: Defaults to true
      responses:
        "200":
          description: The failed or retry-scheduled task
          content:
            application/json:
              schema:
                type: object
                required: [task]
                properties:
                  task:
                    $ref: "#/components/schemas/OctopusTask"
        "400":
          description: Invalid request body
        "404":
          description: Not found
        "409":
          description: Task is not currently claimed by this caller
...
components:
  schemas:
...
    TaskStatus:
      type: string
      enum: [queued, running, completed, failed, cancelled]
    OctopusTask:
      type: object
      description: >
        A unit of work tracked by OS Core's Task Manager + Queue Manager.
        This is the shape every client sees; none of them ever touch the
        `tasks` table directly.
      required:
        - id
        - type
        - queue
        - status
        - priority
        - payload
        - attempts
        - maxAttempts
        - availableAt
        - source
        - correlationId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          description: Dot-namespaced task type, e.g. "campaign.publish"
        queue:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          type: integer
        payload:
          description: Task-type-specific input
        result:
          description: Task-type-specific output, set once completed
          nullable: true
        error:
          type: string
          nullable: true
        attempts:
          type: integer
        maxAttempts:
          type: integer
        availableAt:
          type: string
          format: date-time
        lockedBy:
          type: string
          nullable: true
        lockedAt:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        source:
          type: string
          description: Identity of the enqueuer, e.g. "api-server", "os-core"
        correlationId:
          type: string
          format: uuid
        causationId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTaskRequest:
      type: object
      required: [type, payload]
      properties:
        type:
          type: string
          description: Dot-namespaced task type, e.g. "campaign.publish"
        payload:
          description: Task-type-specific input
        queue:
          type: string
          description: Defaults to "default"
        priority:
          type: integer
          description: Higher is claimed first. Defaults to 0.
        maxAttempts:
          type: integer
          description: Defaults to 3
        delayMs:
          type: integer
          description: Delay before the task becomes claimable. Defaults to 0.
        correlationId:
          type: string
          format: uuid
...